pyproject.toml
[project]
name = "dedupl"
version = "0.2.0"
description = "Media file deduplication toolkit using perceptual fingerprinting"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "deDupl Contributors"}
]
keywords = ["deduplication", "media", "audio", "video", "image", "fingerprinting"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Multimedia :: Video",
    "Topic :: System :: Filesystems",
]

dependencies = [
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
audio = []
image = [
    "Pillow>=9.0.0",
    "imagehash>=4.3.0",
    "pillow-heif>=0.13.0",
]
video = [
    "opencv-python-headless>=4.8.0",
    "Pillow>=9.0.0",
    "imagehash>=4.3.0",
    "numpy>=1.21.0",
]
all = [
    "Pillow>=9.0.0",
    "imagehash>=4.3.0",
    "pillow-heif>=0.13.0",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.21.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]

[project.scripts]
dedupl-audio = "dedupl.cli.audio:main"
dedupl-image = "dedupl.cli.image:main"
dedupl-video = "dedupl.cli.video:main"

[project.urls]
Homepage = "https://github.com/humanrightsconnected/dedupl"
Documentation = "https://github.com/humanrightsconnected/dedupl#readme"
Repository = "https://github.com/humanrightsconnected/dedupl"
Issues = "https://github.com/humanrightsconnected/dedupl/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dedupl"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py38"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
check_untyped_defs = true
no_implicit_reexport = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src/dedupl --cov-report=term-missing"

[tool.coverage.run]
source = ["src/dedupl"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
